Added teams to components and a measure for inter-team communication
The team nodes were added to the dsm, in addition i added a weighting metric for the teams that is the sum of all of the weights of the component interaction for the components that any two teams work on. This represents how much communication with other teams each team has to do. This could still use some work but it gets the point across for now.